+++ src/panels/nav_toolbar.cpp
@@ -54,15 +54,15 @@ NavToolbar::NavToolbar(wxWindow* parent, wxWindowID id, const wxPoint& pos, cons
     AddStretchableSpace();
 
     AddTool(id_NavCollExpand, wxEmptyString,
-        wxueBundleSVG(wxue_img::nav_coll_expand_svg, 291, 562, FromDIP(wxSize(16, 16))), wxNullBitmap, wxITEM_NORMAL,
+        wxueBundleSVG(wxue_img::nav_coll_expand_svg, 291, 562, wxSize(16, 16)), wxNullBitmap, wxITEM_NORMAL,
         "Collapse siblings, expand children", "Expand selected item, collapse all other items at the same level");
 
     AddTool(id_NavExpand, wxEmptyString,
-        wxueBundleSVG(wxue_img::nav_expand_svg, 270, 494, FromDIP(wxSize(16, 16))), wxNullBitmap, wxITEM_NORMAL,
-        "Expand all children", "Expand selected item and all of it\'s sub-items");
+        wxueBundleSVG(wxue_img::nav_expand_svg, 270, 494, wxSize(16, 16)), wxNullBitmap, wxITEM_NORMAL, "Expand all children",
+        "Expand selected item and all of it\'s sub-items");
 
     AddTool(id_NavCollapse, wxEmptyString,
-        wxueBundleSVG(wxue_img::nav_collapse_svg, 214, 340, FromDIP(wxSize(16, 16))), wxNullBitmap, wxITEM_NORMAL,
+        wxueBundleSVG(wxue_img::nav_collapse_svg, 214, 340, wxSize(16, 16)), wxNullBitmap, wxITEM_NORMAL,
         "Collapse all siblings", "Collapse selected item and all items at the same level");
 
     AddSeparator();